- Adapted 64 bit patch from Itzcoalt Alvarez M. <itzcoaltam@joiz.net> - Fixed Russian / Ukraine language mess View Contact

+++ mkahawa-srv/src/EmployeesFrame.h
@@ -30,8 +30,8 @@ protected:
   FXTextField	 *phonetf;
   FXTextField	 *logintf;
 private:
-  int		  editedemployee;
-  int		  usrlvlset;
+  long		  editedemployee;
+  long   	  usrlvlset;
 protected:
   EmployeesFrame(){}
   void drawEmployeeDialog(FXDialogBox *, EmpRecStr *);
