+++ mkahawa-srv/src/ReportFrame.cpp
@@ -858,7 +858,7 @@ long
 ReportFrame::onSessionSelect(FXObject*,FXSelector,void*)
 {
   FXFoldingItem *current = sessionslist->getCurrentItem();
-  int session = (int) current->getData();
+  long session = (long) current->getData();
 
   cashingframe->setSession(session,FALSE);
   mainwin->showCashing();
